+++ site/router.php
@@ -2,14 +2,15 @@
 if (file_exists($_SERVER['DOCUMENT_ROOT'] . $_SERVER['REQUEST_URI'])) {
     return false;
 } else {
-    require '../src/Penibelst/Imadaem/Imadaem.php';
+    require_once '../src/Penibelst/Imadaem/Resizer.php';
+    require_once '../src/Penibelst/Imadaem/Logger.php';
 
-    $log = new Imadaem\Log(E_USER_NOTICE);
-    $imadaem = new Imadaem\Imadaem(
+    $logger = new Penibelst\Imadaem\Logger(E_USER_NOTICE);
+    $resizer = new Penibelst\Imadaem\Resizer(
         array(
             'dstRoot' => 'api',
             'expires' => 7 * 24 * 60 * 60,
             'srcRoot' => 'images'),
-        $log);
-    $imadaem->run();
+        $logger);
+    $resizer->run();
 }
